requirements.txt
#requirements.txt
python-dotenv
numpy
pandas
boto3
paramiko
pymysql
beautifulsoup4
lxml

langchain
langchain-openai
pydantic